Close #507: [`refined4s-pureconfig`] Reorganize type-class instances in the modules
